Utilities Update
- Moved FitToDict to utilities
- AME: added option to pass molecules to get_val method
- MRToFMS: Added simple mass calculator reading from fit files
